Attempted clean slate with rust
This isn't really working and to do this I think I'd have to really
settle down and learn rust, and right now I kind of have too much stuff
swirling in my mind to actually do that and I just want to get this
project done. So with that being said I think I'm going to make this
last commit a new branch and then move to python or ruby.