Final tweaks
Yes, I won't be using Rust but I wanted at least this proof-of-concept
to be in a workable state in Rust. It is "technically" working and if I
kept at this maybe I'd get there but I'd legitimately like to try
Python.

Essentially, in this proof of concept, the number increases every time a
key is pressed because the program halts every time to wait for a key
input. This wouldn't happen with threads and the whole event stream or
poll thingy.