Create template for countdown system
This is essentially the underlying logic under the countdown system, but
currently in terminal format. Eventually, I'll want to move the logic
that actually gets the timedelta between the current datetime and the
target datetime and put that in its own function, and I'll want to make
that return some object of a class that holds the information on the
days, hours, minutes and seconds of the difference. As well as remove
the main function which is currently testing code.